회원 삭제 기능 완료     1. 필드값     이 해당 이메일과 존재하면 삭제 - 닉네임은 사용 안함 x -> schoolScores는     ChatActions, Payments, helpActions : 본인의 이메일이 owner_email, helper_email 에 해당하면 삭제     naver_posts, users : 본인의 이메일이email에 해당하면 삭제     schoolScores: key값이 사용자의 userNickname 인 경우 삭제     naverUserHelpStatus: key값이 사용자의 userEmail인 경우 삭제
    2. 테스트
    - 같은 이메일 중복 회원가입 안 되는 거 다시 확인
    - ChatActions: 상위 컬렉션안 필드는 삭제되었지만 하위 컬렉션 문서(대화내용)들은 삭제 안됨
    - naverUserHelpStatus: 삭제안됨: 키값이 이메일
    - userStatus 문서이름(닉네임)
    - schoolScores: 키값이 닉네임

    - 기존 회원가입은 이메일 인증이 필요하기 때문에 6월 버전으로 회원가입 코드 구성후 테스트
    - Authentication경우 구글로 인증했으면 삭제됨
    - schoolScores, userStatus, naverUserHelpStatus 가 삭제가 안됨
    - naverUserHelpStatus 삭제할 필요 x
    - schoolScores, userStatus -> 아 생성자에서 닉네임하고 이메일을 반대로 받았다.. 해결!

여름 휴가 즉 재정비도 끝났고 로고도 나왔고 달려보자 이지준준

<예정>
안드로이드 폰트 및 햅틱

로딩시 구글 에드몹
랭킹 학기마다 초기화

거래완료시 채팅, 페이, 내역 모두 삭제
출시전 남은거  실제 결제
    이름, 휴대전화 넣어서 페이 테스트(테스트 및 사업자 등록이 필요하면 잠시 대기)
        - accept -> true일때 && 결제가 완료되지 않았을때 (결제 도입 후)
            - 1시간 -> 거래 완료 질문
            - 2시간 -> 한시간 후에 거래 종료 질문
            - 3시간 -> 프로필 개발자에게 묻는 질문에 문제있음 연락 하라고 적는다

        - 거래완료시 채팅, 페이, 내역 모두 삭제

오디오 선정 - 도와주기 전송, 로그인 등등

<프로필 사진> - 출시 전 (왜냐하면 현재는 임의의 계정으로 하기때문에 실제 구글계정이 아니라 프사를 불러올 수가 없다.)
    프로필에서 현재 구글 이미지가 프로필 이미지가 된다 -> 채팅방에서도 적용
    프사 넣기 -> 이거는 구글 계정으로 로그인하면 해당 프로필 사진인데 현재는 내가 임의로 만든 게정이라 안됨

#마지막으로 해야할일
오디오 선정 - 도와주기 전송, 로그인 등등
개인정보 보호규정
앱개발 마케팅 영상
다이소, 편의점도 가능하다는 것을 강조
포스터나 팜플렛 제작시 -> 도와주시 신청과 받는 학생의 장점  / 최소주문을 강조
배달업체에게도 줘보자 -> 한명 정해보기(이건 미정)
개인정보 수집문?

1. 알림 수락시 게시글 삭제, 글작성시 지도 버튼 활성화, 도와주기 텀 제한(30초), 도와주기 요청횟수 제한
2. 다른 학교 추가시
    1. 게시판 수정 - ui
    2. schoolScores 컬렉션 도메인 문서 생성
    3. 학교랭킹 추가 - SchoolRanking.dart,
    4. 초기 중심 위치 설정 (현재 한밭대) - 게시글 작성시
    5. 회원가입시 학교 도메인 추가 - SchoolEmailDialog.dart
    6. 로고사진 스토리지에 넣기
    7. url 파이어스토어에 저장 - updateSchoolLogos()
    8. 바텀 게시판 이동 해당 학교 board 클래스작성
    9. AllUsersScreen.dart, SchoolRanking.dart, Profile.dart, 주문내역에 새로운 학교 import
    10. function index.js -> 대학별 알림 추가
    11. 회원탈퇴시 삭제할때 게시판 이름

3. 테스트 메일 삭제 - gmail.com

+++ lib/Profile/DeleteMember.dart
@@ -1,6 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:flutter/material.dart';
+
 
 class DeleteMember {
   final String userEmail;
@@ -8,12 +8,12 @@ class DeleteMember {
 
   DeleteMember(this.userEmail, this.userNickname);
 
+
   Future<void> deleteUserData() async {
     // 각 컬렉션에서 데이터 삭제
     await _deleteCollectionData('ChatActions');
     await _deleteCollectionData('Payments');
     await _deleteCollectionData('helpActions');
-    await _deleteCollectionData('naverUserHelpStatus');
     await _deleteCollectionData('naver_posts');
     await _deleteCollectionData('schoolScores');
     await _deleteCollectionData('userStatus');
@@ -22,30 +22,84 @@ class DeleteMember {
 
   Future<void> _deleteCollectionData(String collectionName) async {
     try {
-      // 이메일로 데이터 삭제
-      var emailDocs = await FirebaseFirestore.instance
-          .collection(collectionName)
-          .where('email', isEqualTo: userEmail)
-          .get();
+      // 필드 값이 이메일과 닉네임과 일치하는 문서 삭제
+      await _deleteDocumentsByFieldValue(collectionName, userEmail);
+      await _deleteDocumentsByFieldValue(collectionName, userNickname);
 
-      for (var doc in emailDocs.docs) {
-        await FirebaseFirestore.instance.collection(collectionName).doc(doc.id).delete();
-        print("${collectionName} + ${doc.id} + ${userEmail} 삭제완료");
+
+      // 유저 닉네임이 필드값인 경우 삭제
+      if (collectionName == 'schoolScores') {
+        await _deleteDocumentsByFieldKey(collectionName, userNickname);
       }
 
-
-      // 닉네임으로 데이터 삭제
-      var nicknameDocs = await FirebaseFirestore.instance
-          .collection(collectionName)
-          .where('nickname', isEqualTo: userNickname)
-          .get();
-
-      for (var doc in nicknameDocs.docs) {
-        await FirebaseFirestore.instance.collection(collectionName).doc(doc.id).delete();
-        print("${collectionName} + ${doc.id} + ${userNickname} 삭제완료");
-      }
     } catch (e) {
       print('Error deleting data from $collectionName: $e');
+    }
+  }
+
+  Future<void> _deleteDocumentsByFieldValue(String collectionName, String value) async {
+    List<QuerySnapshot> querySnapshots = [];
+
+    // 여러 필드 값으로 문서 삭제
+    if (collectionName == 'ChatActions' || collectionName == 'Payments' || collectionName == 'helpActions') {
+      querySnapshots.add(await FirebaseFirestore.instance.collection(collectionName).where('owner_email', isEqualTo: value).get());
+      querySnapshots.add(await FirebaseFirestore.instance.collection(collectionName).where('helper_email', isEqualTo: value).get());
+
+    } else if (collectionName == 'naver_posts') {
+      querySnapshots.add(await FirebaseFirestore.instance.collection(collectionName).where('email', isEqualTo: value).get());
+    }
+
+    else if (collectionName == 'users' || collectionName == 'userStatus') {
+      await FirebaseFirestore.instance.collection(collectionName).doc(value).delete();
+
+    }
+
+
+    Set<String> deletedDocIds = {};
+    for (var querySnapshot in querySnapshots) {
+      for (var doc in querySnapshot.docs) {
+        if (!deletedDocIds.contains(doc.id)) {
+          //서브컬렉션 삭제
+          if (collectionName == 'ChatActions') {
+            await _deleteSubCollection(doc.reference);
+          }
+
+          await FirebaseFirestore.instance.collection(collectionName).doc(doc.id).delete();
+          deletedDocIds.add(doc.id);
+        }
+      }
+    }
+  }
+
+  Future<void> _deleteDocumentsByFieldKey(String collectionName, String keyToMatch) async {
+    // 해당 컬렉션의 모든 문서를 가져옴
+    QuerySnapshot querySnapshot = await FirebaseFirestore.instance.collection(collectionName).get();
+    // 각 문서의 모든 필드 이름을 검사
+    for (var doc in querySnapshot.docs) {
+      Map<String, dynamic> data = doc.data() as Map<String, dynamic>;
+
+      if (data.containsKey(keyToMatch)) {
+        // 필드 이름이 keyToMatch와 일치하는 경우 해당 필드를 삭제
+        print(keyToMatch);
+        await FirebaseFirestore.instance.collection(collectionName).doc(doc.id).update({
+          keyToMatch: FieldValue.delete()
+        });
+        print('Deleted field $keyToMatch from document with ID: ${doc.id} in collection: $collectionName');
+      }
+    }
+  }
+
+  // ChatActions 서브컬렉션의 모든 문서를 삭제하는 메서드
+  Future<void> _deleteSubCollection(DocumentReference parentDocRef) async {
+    // 서브컬렉션 참조
+    CollectionReference subcollectionRef = parentDocRef.collection('messages');
+
+    // 서브컬렉션의 모든 문서 가져오기
+    QuerySnapshot subcollectionSnapshot = await subcollectionRef.get();
+
+    // 각 문서 삭제
+    for (var doc in subcollectionSnapshot.docs) {
+      await doc.reference.delete();
     }
   }
